Changed: ticks int -> int64

+++ app_apis_paint_text_history.go
@@ -27,7 +27,7 @@ type VmTextHistory struct {
 	items []VmTextHistoryItem
 
 	act          int
-	lastAddTicks int
+	lastAddTicks int64
 }
 
 func NewVmTextHistoryItem(uid *LayoutDiv, init VmTextHistoryItem) *VmTextHistory {
